Update test; change metric names
