+++ instrumentation/spring-cache-3.1.0/src/test/java/com.nr.agent.instrumentation/CacheInstrumentationTest.java
@@ -6,43 +6,101 @@ import com.newrelic.agent.introspec.Introspector;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.springframework.cache.Cache;
+import org.springframework.cache.support.SimpleValueWrapper;
+
+import java.util.HashMap;
 
 @RunWith(InstrumentationTestRunner.class)
 @InstrumentationTestConfig(includePrefixes = { "org.springframework.cache" })
 public class CacheInstrumentationTest {
     @Test
     public void get_withCacheMiss_generatesMissMetric() {
-        SampleCacheImpl cacheImpl = new SampleCacheImpl();
+        Cache cacheImpl = new SampleCacheImpl();
         cacheImpl.get(null);    //Simulate cache miss via null param
 
         Introspector introspector = InstrumentationTestRunner.getIntrospector();
-        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Supportability/Spring/Cache/java.util.HashMap/sample/Misses"));
+        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Spring/Cache/java.util.HashMap/sample/Misses"));
+    }
+
+    @Test
+    public void get_withNullCacheProvider_reportsProviderAsUnknown() {
+        Cache cacheImpl = new SampleCacheWithoutProvider();
+        cacheImpl.get(null);    //Simulate cache miss via null param
+
+        Introspector introspector = InstrumentationTestRunner.getIntrospector();
+        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Spring/Cache/Unknown/sample/Misses"));
     }
 
     @Test
     public void get_withCacheHit_generatesHitMetric() {
-        SampleCacheImpl cacheImpl = new SampleCacheImpl();
+        Cache cacheImpl = new SampleCacheImpl();
         cacheImpl.get("foo");    //Simulate cache hit
 
         Introspector introspector = InstrumentationTestRunner.getIntrospector();
-        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Supportability/Spring/Cache/java.util.HashMap/sample/Hits"));
+        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Spring/Cache/java.util.HashMap/sample/Hits"));
     }
 
     @Test
     public void get_withCacheClear_generatesClearMetric() {
-        SampleCacheImpl cacheImpl = new SampleCacheImpl();
+        Cache cacheImpl = new SampleCacheImpl();
         cacheImpl.clear();
 
         Introspector introspector = InstrumentationTestRunner.getIntrospector();
-        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Supportability/Spring/Cache/java.util.HashMap/sample/Clear"));
+        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Spring/Cache/java.util.HashMap/sample/Clear"));
     }
 
     @Test
     public void get_withCacheEvict_generatesEvictMetric() {
-        SampleCacheImpl cacheImpl = new SampleCacheImpl();
+        Cache cacheImpl = new SampleCacheImpl();
         cacheImpl.evict("foo");
 
         Introspector introspector = InstrumentationTestRunner.getIntrospector();
-        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Supportability/Spring/Cache/java.util.HashMap/sample/Evict"));
+        Assert.assertTrue(introspector.getUnscopedMetrics().containsKey("Spring/Cache/java.util.HashMap/sample/Evict"));
+    }
+
+    private static class SampleCacheImpl implements Cache {
+        @Override
+        public String getName() {
+            return "sample";
+        }
+
+        @Override
+        public Object getNativeCache() {
+            return new HashMap<>();
+        }
+
+        @Override
+        public ValueWrapper get(Object key) {
+            //Simulate miss
+            if (key == null) {
+                return null;
+            }
+
+            //Simulate hit
+            return new SimpleValueWrapper("foo");
+        }
+
+        @Override
+        public void put(Object key, Object value) {
+
+        }
+
+        @Override
+        public void evict(Object key) {
+
+        }
+
+        @Override
+        public void clear() {
+
+        }
+    }
+
+    private static class SampleCacheWithoutProvider extends SampleCacheImpl {
+        @Override
+        public Object getNativeCache() {
+            return null;
+        }
     }
 }
